Feature: Add reference parameter to Ruby SDK
Why is this change necessary?

Payments API currently is taking custom fixed references from
merchants customer create request

How does it address the issue?

It's necessary to test that the custom reference is
not created by Conekta when the customer sends the reference in
a request

What side effects does this have?

Tests related to this feature are skipped because none of the merchants
have offline recurrent identifier which is needed to test this feature